Instead of a long an int is now used for the partition id, saves up some space

+++ hazelblast/src/main/java/com/hazelblast/client/basic/DistributedMethodInvocationFactory.java
@@ -26,5 +26,5 @@ public interface DistributedMethodInvocationFactory {
      * @return the Callable.
      * @throws NullPointerException if sliceName, serviceName, methodName or args is null.
      */
-    <T> Callable<T> create(String sliceName, String serviceName, String methodName, Object[] args, String[] argTypes, long partitionId);
+    <T> Callable<T> create(String sliceName, String serviceName, String methodName, Object[] args, String[] argTypes, int partitionId);
 }
